bitkeeper revision 1.1713.1.8 (42aed659DzIwFCeaK8Cxa8ZpzbHR-Q)
[PATCH] 64-session-id.patch

session_id is an opaque pointer, so should be wider on 64 bit platforms.

+++ tools/xcs/xcs_proto.h
@@ -40,7 +40,7 @@
 #define TYPE_VIRQ      0xfffe
 
 typedef struct {
-    u32 session_id;
+    unsigned long  session_id;
 } xcs_connect_msg_t;
 
 typedef struct {
